fix: renamed IMetricService to metrics.ts
